+++ tests/test_summarize.py
@@ -1,19 +1,12 @@
-from aleph_alpha_client import (
-    AlephAlphaClient,
-    AlephAlphaModel,
-    Document,
-    SummarizationRequest,
-)
+import pytest
+from aleph_alpha_client import Document, SummarizationRequest
 from aleph_alpha_client.aleph_alpha_client import AsyncClient, Client
 
 from tests.common import (
     sync_client,
     async_client,
-    client,
     model_name,
-    luminous_extended,
 )
-import pytest
 
 # AsyncClient
 
@@ -56,25 +49,3 @@ def test_text(sync_client: Client):
 
     assert response.summary is not None
     assert response.model_version is not None
-
-
-# AlephAlphaClient
-
-
-@pytest.mark.system_test
-def test_summarization_with_client(client: AlephAlphaClient):
-    model_name = "luminous-extended"
-    # given a client
-    assert model_name in map(lambda model: model["name"], client.available_models())
-
-    # when posting a Summarization request
-    response = client.summarize(
-        "luminous-extended",
-        SummarizationRequest(
-            document=Document.from_prompt(["Andreas likes pizza."]),
-        ),
-    )
-
-    # The response should exist in the form of a json dict
-    assert response["summary"] is not None
-    assert response["model_version"] is not None